Moved all cluster layer related styling methods to styler service
