+++ components/layermanager/layer-editor-vector-layer.service.ts
@@ -1,31 +1,20 @@
-import Feature from 'ol/Feature';
 import VectorLayer from 'ol/layer/Vector';
-import {Circle, Fill, Stroke, Style, Text} from 'ol/style';
 import {Cluster} from 'ol/source';
 import {HsMapService} from '../map/map.service';
-import {HsQueryVectorService} from '../query/query-vector.service';
+import {HsStylerService} from './../styles/styler.service';
 import {HsUtilsService} from '../utils/utils.service';
 import {Injectable} from '@angular/core';
 import {Layer} from 'ol/layer';
 import {Point} from 'ol/geom';
-import {createDefaultStyle} from 'ol/style/Style';
 
 @Injectable({
   providedIn: 'root',
 })
 export class HsLayerEditorVectorLayerService {
-  clusterStyle = new Style ({
-    stroke: new Stroke({
-      color: '#fff',
-    }),
-    fill: new Fill({
-      color: '#3399CC',
-    }),
-  });
   constructor(
     private HsMapService: HsMapService,
     private HsUtilsService: HsUtilsService,
-    private HsQueryVectorService: HsQueryVectorService
+    private HsStylerService: HsStylerService
   ) {}
 
   /**
@@ -78,7 +67,7 @@ export class HsLayerEditorVectorLayerService {
       layer.set('hsOriginalStyle', layer.getStyle());
       if (!this.HsUtilsService.instOf(layer.getSource(), Cluster)) {
         layer.setSource(this.createClusteredSource(layer, distance));
-        this.styleLayer(layer);
+        this.HsStylerService.styleClusteredLayer(layer);
       }
     } else {
       layer.setStyle(layer.get('hsOriginalStyle'));
@@ -105,121 +94,5 @@ export class HsLayerEditorVectorLayerService {
         }
       },
     });
-  }
-
-  styleLayer(layer: VectorLayer): void {
-    const styleCache = {};
-    layer.setStyle((feature: Feature, resolution: number) => {
-      const size = feature.get('features')?.length || 0;
-      if (size > 1) {
-        return this.makeClusterMarker(styleCache, size);
-      } else {
-        return this.makeSingleFeatureClusterMarker(feature, resolution, layer);
-      }
-    });
-  }
-
-  private makeSingleFeatureClusterMarker(
-    feature: Feature,
-    resolution: number,
-    layer: VectorLayer
-  ) {
-    const featureStyle = feature.get('features')
-      ? feature.get('features')[0].getStyle()
-      : feature.getStyle();
-
-    const originalStyle = featureStyle
-      ? featureStyle
-      : layer.get('hsOriginalStyle')
-      ? layer.get('hsOriginalStyle')
-      : createDefaultStyle;
-
-    let appliedStyle = this.applyStyleIfNeeded(
-      originalStyle,
-      feature,
-      resolution
-    );
-    if (this.isSelectedFeature(feature)) {
-      if (Array.isArray(appliedStyle)) {
-        appliedStyle = appliedStyle[0];
-      }
-      appliedStyle = appliedStyle.clone();
-      appliedStyle.setFill(
-        new Fill({
-          color: [255, 255, 255, 0.5],
-        })
-      );
-      appliedStyle.setStroke(
-        new Stroke({
-          color: [0, 153, 255, 1],
-          width: 3,
-        })
-      );
-    }
-    return this.useStyleOnFirstFeature(appliedStyle, feature);
-  }
-
-  isSelectedFeature(feature: Feature): boolean {
-    if (feature.get('features')) {
-      return this.isSelectedFeature(feature.get('features')[0]);
-    }
-    return (
-      this.HsQueryVectorService.selector
-        .getFeatures()
-        .getArray()
-        .indexOf(feature) > -1
-    );
-  }
-
-  private makeClusterMarker(styleCache: {}, size: any) {
-    let textStyle = styleCache[size];
-    if (!textStyle) {
-      textStyle = new Style({
-        image: new Circle({
-          radius: 10,
-          stroke: this.clusterStyle.getStroke(),
-          fill: this.clusterStyle.getFill(),
-        }),
-        text: new Text({
-          text: size.toString(),
-          fill: new Fill({
-            color: '#000',
-          }),
-        }),
-      });
-      styleCache[size] = textStyle;
-    }
-    return textStyle;
-  }
-
-  private applyStyleIfNeeded(
-    style: any,
-    feature: Feature,
-    resolution: number
-  ): any {
-    if (typeof style == 'function') {
-      return style(feature, resolution);
-    } else {
-      return style;
-    }
-  }
-
-  private useStyleOnFirstFeature(
-    style: any,
-    clusteredContainerFeature: Feature
-  ): Style | Style[] {
-    const originalFeature = clusteredContainerFeature.get('features') || [
-      clusteredContainerFeature,
-    ];
-    let newStyle;
-    if (style.length) {
-      newStyle = style[0].clone();
-      newStyle.setGeometry(originalFeature[0].getGeometry());
-      return [newStyle];
-    } else {
-      newStyle = style.clone();
-      newStyle.setGeometry(originalFeature[0].getGeometry());
-      return newStyle;
-    }
   }
 }
